requirements.txt
asgiref==3.5.2
boto3==1.24.57
botocore==1.27.57
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.0.5
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
django-storages==1.13.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
oauthlib==3.2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
whitenoise==6.2.0
